new models with color and shape
create new levels to the game. this time the complexity increases.
in addition to the 3 colors, we now have a new dimension of shape.
the shape of each object can be either "box" (as before) or "sphere".
for each of the 3 colors, we have the following combinations:
- same color, other shape
- other color 1, same shape
- other color 2, same shape
- other color 1, other shape
- other color 2, other shape

that's 5 x 3 = 15 combinations. plus, we have the same for the other shape.
so we have 30 combinations for the receiver in total.
the target for the sender can see either one of the two choices. so we have
60 combinations in total.

the combinations will appear randomly as before.