adjust notebook for shapes and colors
the LE_notebook.ipynb is now adjusted for shapes as well as colors.
the first 3 bits will be used for colors as usual. next 2 for shapes.
and the last will be reserved for silent utterance as usual.

notebook was executed using the old AE with 50 latent dimensions.
the results look actually good, considering the AE is not trained on
spheres explicitly.

added the results of old notebook with the same AE for comparison.